+++ vllm/model_executor/layers/attention/backends/xformers.py
@@ -53,7 +53,7 @@ class XFormersBackend(nn.Module):
         key_cache: Optional[torch.Tensor],
         value_cache: Optional[torch.Tensor],
         input_metadata: InputMetadata,
-        kv_scale: int,
+        kv_scale: float,
     ) -> torch.Tensor:
         """Forward pass with xFormers and PagedAttention.
